+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-ejb/src/java/org/dcm4chex/archive/ejb/entity/GPPPSBean.java
@@ -83,9 +83,9 @@ public abstract class GPPPSBean implements EntityBean {
      * @jboss.persistence auto-increment="true"
      * 
      */
-    public abstract Integer getPk();
+    public abstract Long getPk();
 
-    public abstract void setPk(Integer pk);
+    public abstract void setPk(Long pk);
 
     /**
      * @ejb.interface-method
@@ -181,7 +181,7 @@ public abstract class GPPPSBean implements EntityBean {
      * 
      * @ejb.create-method
      */
-    public Integer ejbCreate(Dataset ds, PatientLocal patient)
+    public Long ejbCreate(Dataset ds, PatientLocal patient)
             throws CreateException {
         setSopIuid(ds.getString(Tags.SOPInstanceUID));
         setAttributes(ds);
