+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-ejb/src/java/org/dcm4chex/archive/ejb/entity/MWLItemBean.java
@@ -114,9 +114,9 @@ public abstract class MWLItemBean implements EntityBean {
      *  auto-increment="true"
      *
      */
-    public abstract Integer getPk();
+    public abstract Long getPk();
 
-    public abstract void setPk(Integer pk);
+    public abstract void setPk(Long pk);
 
     /**
      * @ejb.interface-method
@@ -269,7 +269,7 @@ public abstract class MWLItemBean implements EntityBean {
      *
      * @ejb.create-method
      */
-    public Integer ejbCreate(Dataset ds, PatientLocal patient) throws CreateException {
+    public Long ejbCreate(Dataset ds, PatientLocal patient) throws CreateException {
         setAttributes(ds);
         return null;
     }
