+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-ejb/src/java/org/dcm4chex/archive/ejb/interfaces/FileSystemDTO.java
@@ -54,7 +54,7 @@ public class FileSystemDTO implements Serializable {
 
     private static final long serialVersionUID = 3257844398468315445L;
 
-    private int pk = -1; // unkown mark
+    private long pk = -1; // unkown mark
 
     private String directoryPath;
 
@@ -85,11 +85,11 @@ public class FileSystemDTO implements Serializable {
         return toString(new StringBuffer()).toString();
     }
 
-    public final int getPk() {
+    public final long getPk() {
         return pk;
     }
 
-    public final void setPk(int pk) {
+    public final void setPk(long pk) {
         this.pk = pk;
     }
